pyproject.toml
[project]
name = "aaa"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bs4>=0.0.2",
    "chromadb>=1.0.20",
    "emoji>=2.14.1",
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "gtts>=2.5.4",
    "langchain-community>=0.3.29",
    "langchain-nomic>=0.1.4",
    "langchain-ollama>=0.3.7",
    "langchain-openai>=0.3.32",
    "nomic[local]>=3.5.3",
    "numpy>=2.3.2",
    "playsound3>=3.2.6",
    "pyaudio>=0.2.14",
    "requests>=2.32.5",
    "sentence-transformers>=5.1.0",
    "sounddevice>=0.5.2",
    "speechrecognition>=3.14.3",
    "streamlit>=1.49.1",
]
